### 2025-12-31 #9716 @editorialbot
[PRE REVIEW]: `tmg-hmc`: A Python package for Exact HMC Sampling for Truncated Multivariate Gaussians with Linear and Quadratic Constraints
**Submitting author:** @erik-a-bensen (<a href="http://orcid.org/0000-0002-2294-1421">Erik A. Bensen</a>)
**Repository:** https://github.com/erik-a-bensen/tmg_hmc
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
