### 2026-01-16 fb8b253
added new section headings along with guidelines to meet the new 2026 JOSS paper guidelines
