### 2026-01-17 5b67917
initial AI usage disclosure
