+++ paper.md
@@ -17,7 +17,7 @@ authors:
 affiliations:
  - name: Department of Statistics and Data Science, Carnegie Mellon University
    index: 1
-date: 1 November 2025
+date: 20 January 2026
 bibliography: paper.bib
 ---
 
@@ -27,10 +27,10 @@ Markov Chain Monte Carlo is a cornerstone of statistical methods that allows us 
  
 # Statement of need
 
- ## Must Include
- * Clearly states what problems the software is designed to solve
- * Identifies who the target audience is
- * Explains the software's relation to other work in the field
+<!-- ## Must Include
+* Clearly states what problems the software is designed to solve
+* Identifies who the target audience is
+* Explains the software's relation to other work in the field -->
 
 Many statistical models of real-world phenomena require the computation of intractable integrals over complex, multivariate probability distributions. Markov Chain Monte Carlo is a foundational statistical method that allows can be used to approximate these quantities from samples [@Robert:1999]. This has allowed for significant progress in statistical modeling in many areas of applied statistics and machine learning [@Gelman:2013]. One important class of distributions that arises due to parameter or data constraints are truncated distributions [@Gelfand:1992; @Swiler:2020; @Stanley:2025]. 
 
@@ -38,9 +38,30 @@ Many statistical models of real-world phenomena require the computation of intra
 $$Q_j(X) \geq 0,\quad\quad j=1,\hdots,m$$,
 where $Q_j(X)$ is a product of linear and quadratic polynomials. As discussed by Pakman and Paninski, this type of a distribution is a critical component of a vast array of statistical models including the probit and Tobit models [@Tobin:1958; @Albert:1993], the dichotomized Gaussian model [@Emrich:1991; @Cox:2002], stochastic integrate-and-fire neural models [@Paninski:2003], Bayesian isotonic regression [@Neelon:2004], and the Bayesian bridge model [@Polson:2014].
 
-More recently, distributions of this form have been used for learning partially censored Gaussian processes [@Cao:2025]. The Exact HMC algorithm was used to implement physics-informed constraints for fields governing CO2 flux in the WOMBAT v2.0 hierarchical flux-inversion framework for inferring changes to the global carbon cycle [@Bertolacci:2024]. Additionally, sampling constrained multivariate normals is relevant to a growing range of literature on constrained Gaussian processes [@Bachoc:2019; @Bachoc:2022; @Swiler:2020; @Agrell:2019; @DaVeiga:2012; @Lopez:2018; @Maatouk:2017]. In particular, in our ongoing research, we are using `tmg-hmc` to sample random transport maps given by the gradient of 2d convex Gaussian processes. We do this by approximating the convex GP by imposing quadratic convex inequality constraints on a discrete spatial grid. \autoref{fig:tmap} shows an example of such a transport map sampled using `tmg-hmc`.
+More recently, distributions of this form have been used for learning partially censored Gaussian processes [@Cao:2025]. The Exact HMC algorithm was used to implement physics-informed constraints for fields governing CO2 flux in the WOMBAT v2.0 hierarchical flux-inversion framework for inferring changes to the global carbon cycle [@Bertolacci:2024]. Additionally, sampling constrained multivariate normals is relevant to a growing range of literature on constrained Gaussian processes [@Bachoc:2019; @Bachoc:2022; @Swiler:2020; @Agrell:2019; @DaVeiga:2012; @Lopez:2018; @Maatouk:2017]. 
+
+# Research Impact Statement
+<!-- The paper must provide evidence of either
+
+## Realized Impact
+* Publications using the software
+* Evidence of external use and adoption
+* Integration with other research tools or workflows status
+
+## OR credible near-term significance
+* Benchmark results demonstrating improvements
[…]
