### 2026-01-20 b763cf6
fixed minor equation formatting issue
