### 2026-02-28 859fccf
finalized paper edits
