+++ README.md
@@ -195,7 +195,7 @@ This project is licensed under the MIT License - see the [LICENSE](LICENSE) file
 If you use this package in your research, please cite:
 
 **Software:**
-> Bensen, E. A., & Kuusela, M. (2025). tmg_hmc: A Python package for Exact HMC Sampling for Truncated Multivariate Gaussians with Linear and Quadratic Constraints. *TBD*. [To be updated upon acceptance]
+> Bensen, E. A., & Kuusela, M. (2026). tmg_hmc: A Python package for Exact HMC Sampling for Truncated Multivariate Gaussians with Linear and Quadratic Constraints. *Journal of Open Source Software*. [In Review]
 
 **Methodology:**
 > Pakman, A., & Paninski, L. (2014). Exact Hamiltonian Monte Carlo for Truncated Multivariate Gaussians. *Journal of Computational and Graphical Statistics*, 23(2), 518-542. https://doi.org/10.1080/10618600.2013.788448
@@ -204,12 +204,12 @@ If you use this package in your research, please cite:
 <summary>BibTeX</summary>
 
 ```bibtex
-@article{Bensen2025tmghmc,
+@article{Bensen2026tmghmc,
   title={tmg\_hmc: A Python package for Exact HMC Sampling for Truncated Multivariate Gaussians with Linear and Quadratic Constraints},
   author={Bensen, Erik A. and Kuusela M.},
-  journal={TBD},
-  year={2025},
-  note={[To be updated upon acceptance]}
+  journal={Journal of Open Source Software},
+  year={2026},
+  note={[In Review]}
 }
 
 @article{PakmanPaninski2014,
