### 2026-02-28 50065bc
fixed bibtex issue
