### 2026-02-28 f78e446
Add packaging requirement to build system
