+++ pyproject.toml
@@ -1,6 +1,7 @@
 [build-system]
 requires = [
     "setuptools>=61.0",
+    "packaging>=24.2",
     "wheel",
     "pybind11>=2.10"
 ]
