### 2026-02-28 449902c
bump version to 1.0.2
