### 2026-02-28 a30d1df
Update API documentation [skip ci]
