### 2026-02-28 a2e54f5
Change license format in pyproject.toml
