### 2026-02-28 0f36a42
Remove packaging requirement from pyproject.toml
