+++ pyproject.toml
@@ -1,7 +1,6 @@
 [build-system]
 requires = [
     "setuptools>=61.0",
-    "packaging>=24.2",
     "wheel",
     "pybind11>=2.10"
 ]
