### 2026-02-28 b121359
Update API documentation [skip ci]
