### 2026-03-27 #9716 @crvernon
@erik-a-bensen please reduce the word count of your paper to less than 1800 words.  You can move some content to your docs and point to them from the paper if needed.  Just as long as the required sections stay in the paper.

Also, please remove any formatting (e.g., backticks) from your paper title as this will mess up JOSS formatting if your paper is accepted for publication.

Thank you!
