+++ paper.md
@@ -27,11 +27,6 @@ Markov Chain Monte Carlo is a cornerstone of statistical methods that allows us 
  
 # Statement of need
 
-<!-- ## Must Include
-* Clearly states what problems the software is designed to solve
-* Identifies who the target audience is
-* Explains the software's relation to other work in the field -->
-
 Many statistical models of real-world phenomena require the computation of intractable integrals over complex, multivariate probability distributions. Markov Chain Monte Carlo is a foundational statistical method that can be used to approximate these quantities from samples [@Robert:1999]. This has allowed for significant progress in statistical modeling in many areas of applied statistics and machine learning [@Gelman:2013]. One important class of distributions that arises due to parameter or data constraints are truncated distributions [@Gelfand:1992; @Swiler:2020; @Stanley:2025]. 
 
 In @Pakman:2014, Exact HMC is used to sample from a $d$-dimensional Gaussian $X \sim N(\mu,\Sigma)$ that is truncated with $m$ inequality constraints of the form 
@@ -42,20 +37,6 @@ More recently, distributions of this form have been used for learning partially 
 
 # Research Impact Statement
 
-<!-- The paper must provide evidence of either
-
-## Realized Impact
-* Publications using the software
-* Evidence of external use and adoption
-* Integration with other research tools or workflows status
-
-## OR credible near-term significance
-* Benchmark results demonstrating improvements
-* Reproducible research materials showing capabilities
-* Community-readiness signals (e.g. requests from other groups, presentations at relevant venues)
-The evidence should be compelling and specific, not aspirational. -->
-
-
 In our ongoing statistical research, we are using `tmg-hmc` to sample random transport maps given by the gradient of 2d convex Gaussian processes. We do this by approximating the convex GP by imposing quadratic convex inequality constraints on a discrete spatial grid. \autoref{fig:tmap} shows an example of such a transport map sampled using `tmg-hmc`. For context, one constraint is positivity of the second derivative for $416$ coordinates. If we were to naively rejection sample with just these positivity constraints for a GP with small length scales so that the locations are approximately independent, the acceptance probability would approach $1/2^{416}$ which is completely infeasible. We also tested sampling by using a linearly constrained region as a proposal distribution and then rejection sampling to get to the convexity constraint, in this case we get an acceptance propability of $\approx1/10^4$ which is improved over naive rejection sampling, but still infeasible for most applications. `tmg-hmc` allows us to sample directly from the convex Gaussian process distribution and we are able to sample several thousand samples per second.
 
 ![Sample of a random transport map defined as the gradient of a 2d convex Gaussian process.\label{fig:tmap}](./resources/example_tmap.png){ width=50% }
@@ -64,80 +45,15 @@ Preliminary work, including our use of `tmg-hmc`, has already been presented at 
 
 # State of the Field
 
-<!-- Authors must describe how their software compares to other commonly-used packages in the research area. When related tools exist, authors must provide a clear ``build vs. contribute'' justification explaining
-* Their unique scholarly contribution
-* Why existing alternatives are insufficient for their research needs
-* What gap their software fills -->
-
[…]
