### 2026-03-30 #9716 @logological
@crvernon: Thanks for the invitation, but I won't be able to take on this manuscript at this time.
