### 2026-03-30 #9716 @matt-graham
Hi @crvernon if it's okay with you I am happy to be the editor for this submission; I was going to self-nominate earlier but then I spotted your message to logological so wanted to wait for their response first.
