### 2026-03-31 #9716 @matt-graham
👋 @SuperKam91 @AdrienCorenflos @dirmeier, can you review this submission for JOSS?
