### 2026-04-29 #10405 @AdrienCorenflos
At this stage, I would have a few questions/comments for @erik-a-bensen to consider

1. Journal guidelines ask to justify the contribution of each authors to the repository. At this time, the contribution of the second author is not clear: no commit by them is recorded to the library under review.
2. The overall impact of the library is not clear. While there is an impact statement, it is essentially limited to saying that the authors have been using it for a related project. I do not believe this, or the current usage of the library, are enough to satisfy JOSS requirement.
3. The implemented method is arguably not the state of the art in MCMC for truncated Gaussians, see https://www.tandfonline.com/doi/full/10.1080/01621459.2024.2395587 for a discussion. It is not implausible that the library can be extended/modified to allow for different implementations and comparison/etc, but as it stands, the sole method implemented therein is outdated.
