### 2026-05-01 #10405 @matt-graham
Thanks @AdrienCorenflos for your comments.

@erik-a-bensen to follow up on Adrien's point around the author list, the relevant [JOSS guidance on authorship](https://joss.readthedocs.io/en/latest/submitting.html#authorship) is

> Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are. The authors themselves assume responsibility for deciding who should be credited with co-authorship, and co-authors must always agree to be listed.

So we need to know here if you and your co-author, Mikael Kuusela, think they qualifies for co-authorship under that definition.

> The implemented method is arguably not the state of the art in MCMC for truncated Gaussians...

With regards to this point (3) @AdrienCorenflos -  just to check was this comment a request for the authors to expand their discussion of related algorithms in the 'State of the field' section of the paper to include Hamiltonian zigzag samplers? For clarification, there is no specific requirement for JOSS submissions to implement 'state of the art' algorithms, only that the software has a research application and that any performance claims made are verifiable / reproducible. However, it does look like the [hdtg package](https://cran.r-project.org/web/packages/hdtg/index.html) used and cited in the paper @AdrienCorenflos  mentioned would be a useful addition to include in your  'State of the field' section @erik-a-bensen as it appears to implement both the 'exact HMC' algorithm of Pakman and Paninski (2014) provided by `tmg-hmc` and the zigzag HMC algorithm.
