### 2026-05-01 #10405 @AdrienCorenflos
@matt-graham to be more precise, the main issue I have is not necessarily the fact that the state of the art is *not* implemented, but rather that I do not believe the current implementation of the package is extensible enough to accommodate it (or another one) if necessary. As it stands, the library tackles the implementation of a specific algorithm, rather than solves a specific problem.
