### 2026-05-10 #10405 @erik-a-bensen
Thank you for your comments @AdrienCorenflos and @matt-graham, below are in-line responses.

> Journal guidelines ask to justify the contribution of each authors to the repository. At this time, the contribution of the second author is not clear: no commit by them is recorded to the library under review.

- My co-author, Mikael Kuusela (@mkuusela), is my PhD advisor and he has been extensively involved in the development of this project. This involved active project direction through regular meetings to give guidance on package design, working through implementation challenges, speeding up the sampling process, as well as providing feedback and revisions on the JOSS paper. Both Mikael and I believe that this more than satisfies the requirements for co-authorship under the JOSS guidelines.

> The overall impact of the library is not clear. While there is an impact statement, it is essentially limited to saying that the authors have been using it for a related project. I do not believe this, or the current usage of the library, are enough to satisfy JOSS requirement.

- When we were preparing this package for submission, we based the impact statement on the following guidelines.
     - JOSS author guidelines: “There must be evidence that the software is being used for research — at minimum by the developers themselves, and ideally by others. Acceptable signals include: references in published papers or preprints, DOIs linking to the software”
     - JOSS reviewer guidelines: “The paper must provide evidence of either:
          - Realized Impact
                    - Publications using the software
          - Credible near-term significance
                    - Reproducible research materials showing capabilities
                    - Community-readiness signals e.g. presentations at relevant venues
- I do believe that we satisfy these guidelines
     - Publications using the software:
        - As mentioned in the manuscript, we are in the final stages of preparing a manuscript that will be submitted to the Annals of Applied Statistics that relies critically on, and will be citing this package.
     - Reproducible research materials showing capabilities:
         - We believe that this is demonstrated by the random transport maps we discuss in the JOSS submission. These transport maps are defined as gradients of a 2D Gaussian process and are sampled as truncated Gaussians with quadratic inequality constraints. As we discuss in this section, approximating the quadratic constraints with linear constraints leads to unsatisfactory acceptance probabilities. Therefore the alternative methods we discuss, and that @AdrienCorenflos discusses in the next comment, which only allow for linearly truncated Gaussians are not viable.
         - Additionally, our work is done using Python, and there is a growing literature of statistical research that uses Python and it still remains that this implementation, to the best of our knowledge, is the only existing Python implementation of exact HMC for truncated Gaussians. I am happy to edit the wording of the statement of need to be more explicit about this.
     - Community-readiness signals:
         - As we mention in the statement, we have presented our work using this package at several statistics and geophysics conferences and received great interest from the community. Some of these presentations for which resources are publicly available include:
            - Minisymposium at the SIAM Uncertainty Quantification 2026 conference: https://stat.cmu.edu/~mkuusela/Kuusela_SIAM_UQ26.pdf (slide 14 cites the JOSS paper)
            - STAMPS Workshop on Neural Simulation Based Inference 2025: https://www.cmu.edu/dietrich/statistics-datascience/stamps/images/bensen_poster_sbi_workshop_2025.pdf
            - Joint Statistical Meeting 2025, invited topics session on Inference Without Exact Likelihooods: https://ww3.aievolution.com/JSMAnnual2025/Events/viewEv?ev=2226
         - We don’t mention the specific conferences or presentations in the paper as it is not customary to explicitly cite presentations or posters and we were asked during the pre-review period to keep this paper below 1800 words. If you would like us to explicitly cite these materials to demonstrate community readiness, please let me know how you would prefer us to do this.

> The implemented method is arguably not the state of the art in MCMC for truncated Gaussians, see https://www.tandfonline.com/doi/full/10.1080/01621459.2024.2395587 for a discussion. It is not implausible that the library can be extended/modified to allow for different implementations and comparison/etc, but as it stands, the sole method implemented therein is outdated.

- Respectfully, I would like to push back on the comment that the Exact HMC / Harmonic HMC method is not state-of-the-art and that it is outdated.  I am interpreting your comments as arising from the claims in Section 4 of the Zigzag paper that: “In fact, the numerical results of Zhang et al. (2022) indicate zigzag HMC as a preferred choice over the algorithms of Pakman and Paninski (2014) and Botev (2017) in many high-dimensional applications.” as well as their truncated Gaussian experimental results.
- In the Zhang et al. (2022) paper that introduces the hdtg package, the authors define a truncated multivariate Gaussian as a random vector x such that  $$x\sim N(\mu,\Sigma)\text{ with }l \leq x \leq u\text{ bounded}$$
This is only a linearly constrained Gaussian which is a much more restrictive set of distributions than those of the Exact HMC algorithm of Pakman and Paninsky (2014) that can handle both linear and quadratic inequality constraints as well as arbitrary products of linear and quadratic inequality constraints. To the best of our knowledge, this is still the only method that is used for this setting, and this is the setting we set about addressing when we developed this package.
- In the Zigzag paper you cite, they also only use linearly constrained Gaussians with the Threshold Model posteriors in section 4.3 being constrained to the positive orthant, and the Phylogenetic Tree in section 4.4 uses a Gaussian truncated to $\{\text{sign}(x)=y\}$ for $y\in\{1,-1\}^d$ which is another linear truncation.
- Additionally, if you look at the simulation results from Table 1 of the Zhang et. al. (2022) paper https://arxiv.org/pdf/2210.01097 you see that even in the linear truncation setting, Harmonic HMC (Exact HMC) performs very well. In the three models tested, Harmonic HMC substantially outperforms zigzag in all tested dimensions for the CS0.9 model, and in all except d=1600 for the LKJ model. The zigzag sampler performs better in the HIV model; however, the only dimensions where this appears substantial is for d=800,1600. The authors even state that there are many settings where Harmonic HMC is the preferred choice. I think this is clear evidence that harmonic HMC is far from outdated and  is clearly still highly useful.
- While it may be plausible that the zigzag methods can be extended to the quadratic constraint setting, it has not been done and a development of this sort would be an involved statistical methods development that is beyond the scope of a JOSS submission. For this reason, I do not believe it is realistic to expect us to make this adaptation to our codebase for this submission.
- This all being said, I am happy to extend the state of the field section to include a discussion of the hdtg package with the acknowledgment that like the Vecchia and Nearest Neighbor approximate samplers, it is written only for R and covers a more restrictive setting of distributions than our Exact HMC implementation.
- We would also like to clarify that we're not intending this to be a generic sampler library. As such, we do not foresee extending the library to include other samplers. Instead the goal is to provide an efficient and well-designed implementation of Exact HMC since this algorithm is highly useful in a variety of contexts and no such implementations exist in Python. The kinds of extensions one could consider are extensions beyond Gaussian distributions or to more general constraints and the library is designed to be extensible in this way.
