### 2026-05-21 2804e88
updated pyproject .toml and README to have an examples option for installation
