### 2026-05-21 a21a5c0
added initial CONTRIBUTING.md with instructions on using ruff for formatting and style
