### 2026-05-21 37a6906
added more documentation about how to access example notebooks
