### 2026-05-21 882f3c0
initial ruff runthrough and fixes for /src
