+++ src/tmg_hmc/gpu_utils.py
@@ -1,0 +1,14 @@
+import importlib.util
+
+_TORCH_AVAILABLE = importlib.util.find_spec("torch") is not None
+
+if _TORCH_AVAILABLE:
+    import torch
+    Tensor = torch.Tensor
+else:
+    class Tensor:
+        pass
+
+    class torch:
+        Tensor = Tensor
+        sparse_coo = None
