+++ src/tmg_hmc/quad_solns.py
@@ -1,5 +1,5 @@
 from cmath import sqrt
-from tmg_hmc.utils import arccos 
+from tmg_hmc.utils import arccos
 
 def soln1(q1: float, q2: float, q3: float, q4: float, q5: float) -> float:
     """
@@ -27,7 +27,7 @@ def soln1(q1: float, q2: float, q3: float, q4: float, q5: float) -> float:
     -----
     DO NOT MODIFY THIS FUNCTION
     The solution is derived from the quartic equation associated with the quadratic constraint hit time
-    given in Eqns 2.48 - 2.53 of Pakman and Paninski (2014). The expression is computed exactly using 
+    given in Eqns 2.48 - 2.53 of Pakman and Paninski (2014). The expression is computed exactly using
     Mathematica and then exported to Fortran which uses the same syntax as Python for mathematical operations.
     See resources/HMC_exact_soln.nb for details.
 
@@ -35,86 +35,86 @@ def soln1(q1: float, q2: float, q3: float, q4: float, q5: float) -> float:
     use the compiled module for efficient computation of all solutions.
     This function is maintained for reference and validation purposes.
     """
-    return (-arccos(-0.5*(q1*q2 + q4*q5)/(q1**2 + q4**2) - 
-         sqrt((q1*q2 + q4*q5)**2/(q1**2 + q4**2)**2 - 
-            (2*(q2**2 + 2*q1*q3 - q4**2 + q5**2))/(3.*(q1**2 + q4**2)) + 
-            (2**(1./3)*(-12*(q2*q3 - q4*q5)*(q1*q2 + q4*q5) + 
+    return (-arccos(-0.5*(q1*q2 + q4*q5)/(q1**2 + q4**2) -
+         sqrt((q1*q2 + q4*q5)**2/(q1**2 + q4**2)**2 -
+            (2*(q2**2 + 2*q1*q3 - q4**2 + q5**2))/(3.*(q1**2 + q4**2)) +
+            (2**(1./3)*(-12*(q2*q3 - q4*q5)*(q1*q2 + q4*q5) +
                  12*(q1**2 + q4**2)*(q3**2 - q5**2) + (q2**2 + 2*q1*q3 - q4**2 + q5**2)**2))/
-             (3.*(q1**2 + q4**2)*(108*(q1**2 + q4**2)*(q2*q3 - q4*q5)**2 + 
-                  108*(q1*q2 + q4*q5)**2*(q3**2 - q5**2) - 
-                  36*(q2*q3 - q4*q5)*(q1*q2 + q4*q5)*(q2**2 + 2*q1*q3 - q4**2 + q5**2) - 
-                  72*(q1**2 + q4**2)*(q3**2 - q5**2)*(q2**2 + 2*q1*q3 - q4**2 + q5**2) + 
-                  2*(q2**2 + 2*q1*q3 - q4**2 + q5**2)**3 + 
-                  sqrt(-4*(-12*(q2*q3 - q4*q5)*(q1*q2 + q4*q5) + 12*(q1**2 + q4**2)*(q3**2 - q5**2) + 
[…]
