+++ src/tmg_hmc/constraints.py
@@ -1,13 +1,11 @@
 from __future__ import annotations
 import numpy as np
-from typing import Protocol, Tuple, Protocol
+from typing import Tuple, Protocol
 from tmg_hmc.utils import Array, Sparse, to_scalar, get_sparse_elements
 from tmg_hmc.quad_solns import soln1, soln2, soln3, soln4, soln5, soln6, soln7, soln8
-from tmg_hmc import get_torch, get_tensor_type
+from tmg_hmc.gpu_utils import torch, Tensor, _TORCH_AVAILABLE
 from tmg_hmc.compiled import calc_all_solutions
 import warnings
-
-torch, Tensor = get_torch(), get_tensor_type()
 
 pis = np.array([-1, 0, 1]) * np.pi
 eps = 1e-12
@@ -35,7 +33,7 @@ class Constraint(Protocol):
         bool
             True if the constraint is satisfied, False otherwise
         """
-        return self.value(x) >= 0 
+        return self.value(x) >= 0
 
     def is_zero(self, x: Array) -> Tuple[bool, bool]:
         """
@@ -52,7 +50,7 @@ class Constraint(Protocol):
         """
         val = self.value(x)
         return np.isclose(val, 0), np.isclose(val, 0, atol=1e-2)
-    
+
     def compute_q(self, a: Array, b: Array) -> Tuple[float, ...]:
         """
         Compute the coefficients of the constraint equation along the trajectory defined by a and b
@@ -105,7 +103,7 @@ class Constraint(Protocol):
             Dictionary representation of the constraint
         """
         d = self.__dict__.copy()
-        
+
         # For sparse constraints, ensure we save S directly
         # and remove the individual row/column vectors that cause reconstruction issues
         if 'sparse' in d and d['sparse']:
@@ -117,15 +115,15 @@ class Constraint(Protocol):
             for key in keys_to_remove:
                 if key in d:
                     del d[key]
-        
+
         # Convert tensors to CPU
         for k, v in d.items():
             if isinstance(v, Tensor):
                 d[k] = v.cpu()
-        
+
         d['type'] = self.__class__.__name__
         return d
-    
+
     @classmethod
     def deserialize(cls, d: dict, gpu: bool) -> Constraint:
         """
@@ -155,7 +153,7 @@ class Constraint(Protocol):
             return QuadraticConstraint.build_from_dict(d, gpu)
         else:
             raise ValueError(f"Unknown constraint type {d['type']}")
-    
+
 
 class ProductConstraint(Constraint):
     """
@@ -188,7 +186,7 @@ class ProductConstraint(Constraint):
         for constraint in self.constraints:
             val *= constraint.value(x)
         return val
-    
+
     def normal(self, x: Array) -> Array:
         """
         Compute the normal vector of the product constraint at x
@@ -207,7 +205,7 @@ class ProductConstraint(Constraint):
         normals = [c.normal(x) for c in self.constraints]
         weighted = [normals[i] * np.prod(vals[:i] + vals[i+1:]) for i in range(len(self.constraints))]
         return sum(weighted)
-    
+
     def hit_time(self, x: Array, xdot: Array) -> Array:
         """
         Compute the hit time of the product constraint along the trajectory defined by x and xdot
@@ -236,7 +234,7 @@ class LinearConstraint(Constraint):
     Constraint of the form fx + c >= 0
     """
     def __init__(self, f: Array, c: float) -> None:
-        """ 
+        """
         Parameters
         ----------
         f : Array
@@ -246,7 +244,7 @@ class LinearConstraint(Constraint):
         """
         self.f = f
         self.c = c
-    
+
     def value(self, x: Array) -> float:
         """
         Compute the value of the constraint at x
[…]
