### 2026-05-21 c188503
initial ruff runthrough and fixes for tests/
