+++ tests/test_utils.py
@@ -1,4 +1,3 @@
-import pytest 
 import numpy as np
 from hypothesis import given, strategies as st
 
