### 2026-05-21 167fa5c
initial ruff format of src files
