+++ src/tmg_hmc/quad_solns.py
@@ -1,6 +1,7 @@
 from cmath import sqrt
 from tmg_hmc.utils import arccos
 
+
 def soln1(q1: float, q2: float, q3: float, q4: float, q5: float) -> float:
     """
     Computes the first of the 8 solutions for the full quadratic constraint hit time.
@@ -35,88 +36,295 @@ def soln1(q1: float, q2: float, q3: float, q4: float, q5: float) -> float:
     use the compiled module for efficient computation of all solutions.
     This function is maintained for reference and validation purposes.
     """
-    return (-arccos(-0.5*(q1*q2 + q4*q5)/(q1**2 + q4**2) -
-         sqrt((q1*q2 + q4*q5)**2/(q1**2 + q4**2)**2 -
-            (2*(q2**2 + 2*q1*q3 - q4**2 + q5**2))/(3.*(q1**2 + q4**2)) +
-            (2**(1./3)*(-12*(q2*q3 - q4*q5)*(q1*q2 + q4*q5) +
-                 12*(q1**2 + q4**2)*(q3**2 - q5**2) + (q2**2 + 2*q1*q3 - q4**2 + q5**2)**2))/
-             (3.*(q1**2 + q4**2)*(108*(q1**2 + q4**2)*(q2*q3 - q4*q5)**2 +
-                  108*(q1*q2 + q4*q5)**2*(q3**2 - q5**2) -
-                  36*(q2*q3 - q4*q5)*(q1*q2 + q4*q5)*(q2**2 + 2*q1*q3 - q4**2 + q5**2) -
-                  72*(q1**2 + q4**2)*(q3**2 - q5**2)*(q2**2 + 2*q1*q3 - q4**2 + q5**2) +
-                  2*(q2**2 + 2*q1*q3 - q4**2 + q5**2)**3 +
-                  sqrt(-4*(-12*(q2*q3 - q4*q5)*(q1*q2 + q4*q5) + 12*(q1**2 + q4**2)*(q3**2 - q5**2) +
-                        (q2**2 + 2*q1*q3 - q4**2 + q5**2)**2)**3 +
-                    (108*(q1**2 + q4**2)*(q2*q3 - q4*q5)**2 + 108*(q1*q2 + q4*q5)**2*(q3**2 - q5**2) -
-                       36*(q2*q3 - q4*q5)*(q1*q2 + q4*q5)*(q2**2 + 2*q1*q3 - q4**2 + q5**2) -
-                       72*(q1**2 + q4**2)*(q3**2 - q5**2)*(q2**2 + 2*q1*q3 - q4**2 + q5**2) +
-                       2*(q2**2 + 2*q1*q3 - q4**2 + q5**2)**3)**2))**(1./3)) +
-            (108*(q1**2 + q4**2)*(q2*q3 - q4*q5)**2 + 108*(q1*q2 + q4*q5)**2*(q3**2 - q5**2) -
-                36*(q2*q3 - q4*q5)*(q1*q2 + q4*q5)*(q2**2 + 2*q1*q3 - q4**2 + q5**2) -
-                72*(q1**2 + q4**2)*(q3**2 - q5**2)*(q2**2 + 2*q1*q3 - q4**2 + q5**2) +
-                2*(q2**2 + 2*q1*q3 - q4**2 + q5**2)**3 +
-                sqrt(-4*(-12*(q2*q3 - q4*q5)*(q1*q2 + q4*q5) + 12*(q1**2 + q4**2)*(q3**2 - q5**2) +
-                      (q2**2 + 2*q1*q3 - q4**2 + q5**2)**2)**3 +
[…]
