+++ src/tmg_hmc/sampler.py
@@ -1,18 +1,34 @@
 from __future__ import annotations
 import numpy as np
 from typing import Tuple
-from tmg_hmc.constraints import Constraint, LinearConstraint, SimpleQuadraticConstraint, QuadraticConstraint, ProductConstraint
+from tmg_hmc.constraints import (
+    Constraint,
+    LinearConstraint,
+    SimpleQuadraticConstraint,
+    QuadraticConstraint,
+    ProductConstraint,
+)
 from tmg_hmc.utils import Array, sparsify, is_nonzero_array
 import warnings
 import pickle
 from tmg_hmc.gpu_utils import torch
+
 
 class TMGSampler:
     """
     Hamiltonian Monte Carlo sampler for Multivariate Gaussian distributions
     with linear and quadratic constraints.
     """
-    def __init__(self, mu: Array = None, Sigma: Array = None, T: float = np.pi/2, gpu: bool = False,*,Sigma_half: Array = None) -> None:
+
+    def __init__(
+        self,
+        mu: Array = None,
+        Sigma: Array = None,
+        T: float = np.pi / 2,
+        gpu: bool = False,
+        *,
+        Sigma_half: Array = None,
+    ) -> None:
         """
         Parameters
         ----------
@@ -87,7 +103,7 @@ class TMGSampler:
         if not all_positive:
             min_eig = torch.min(s) if self.gpu else np.min(s)
             if abs(min_eig) < 1e-10:
-                s -= 2*min_eig
+                s -= 2 * min_eig
             else:
                 raise ValueError("Sigma must be positive semi-definite")
         if self.gpu:
@@ -123,7 +139,7 @@ class TMGSampler:
         if not all_positive:
             min_eig = torch.min(s) if self.gpu else np.min(s)
             if abs(min_eig) < 1e-10:
-                s -= 2*min_eig
+                s -= 2 * min_eig
             else:
                 raise ValueError("Sigma_half must be positive semi-definite")
         if self.gpu:
@@ -131,7 +147,15 @@ class TMGSampler:
         else:
             self.Sigma_half = V @ np.diag(s) @ V.T
 
-    def _build_constraint(self, *, A: Array = None, f: Array = None, c: float = 0.0, sparse: bool = True, compiled: bool = True) -> Constraint:
+    def _build_constraint(
+        self,
+        *,
+        A: Array = None,
+        f: Array = None,
+        c: float = 0.0,
+        sparse: bool = True,
+        compiled: bool = True,
+    ) -> Constraint:
         """
         Builds a constraint to the sampler of the form:
             x.T @ A @ x + f.T @ x + c >= 0
@@ -188,9 +212,9 @@ class TMGSampler:
             c_new = c + mu.T @ Amu + mu.T @ f
         elif (A is not None) and (f is None):
             Amu = A @ mu
-            #f_new = 2*S @ A @ mu
+            # f_new = 2*S @ A @ mu
             f_new = S @ Amu * 2
-            #c_new = c + mu.T @ A @ mu
+            # c_new = c + mu.T @ A @ mu
             c_new = mu.T @ Amu + c
         elif (A is None) and (f is not None):
             f_new = S @ f
@@ -205,7 +229,7 @@ class TMGSampler:
         if self.gpu:
             c_new = c_new.item()
         else:
-            c_new = c_new[0,0]
+            c_new = c_new[0, 0]
 
         if nonzero_A and nonzero_f:
             return QuadraticConstraint(A, f_new, c_new, S, sparse, compiled)
@@ -214,9 +238,19 @@ class TMGSampler:
         elif (not nonzero_A) and nonzero_f:
             return LinearConstraint(f_new, c_new)
         else:
[…]
