+++ src/tmg_hmc/utils.py
@@ -7,11 +7,12 @@ import os
 from tmg_hmc.gpu_utils import torch, Tensor
 
 # ignore runtime warning
-np.seterr(divide='ignore', invalid='ignore')
+np.seterr(divide="ignore", invalid="ignore")
 
 Array: TypeAlias = np.ndarray | Tensor | coo_matrix | None
 Sparse: TypeAlias = csc_matrix | csr_matrix | coo_matrix
 base_path = os.path.dirname(os.path.abspath(__file__))
+
 
 def compiled_library_available() -> bool:
     """
@@ -23,6 +24,7 @@ def compiled_library_available() -> bool:
         True if the shared library is available, False otherwise.
     """
     return importlib.util.find_spec("tmg_hmc.compiled") is not None
+
 
 def sparsify(A: Array) -> Array:
     """
@@ -44,6 +46,7 @@ def sparsify(A: Array) -> Array:
         return A.to_sparse()
     else:
         raise ValueError(f"Unknown type {type(A)}")
+
 
 def get_sparse_elements(A: Array) -> Tuple[Array, Array, Array]:
     """
@@ -74,6 +77,7 @@ def get_sparse_elements(A: Array) -> Tuple[Array, Array, Array]:
     else:
         raise ValueError(f"Unknown type {type(A)}")
 
+
 def to_scalar(x: Array | float) -> float:
     """
     Converts a scalar array or a float to a float.
@@ -94,7 +98,8 @@ def to_scalar(x: Array | float) -> float:
         return x.cpu().item()
     elif len(x.shape) == 1:
         return x[0]
-    return x[0,0]
+    return x[0, 0]
+
 
 def is_nonzero_array(x: Array) -> bool:
     """
@@ -119,6 +124,7 @@ def is_nonzero_array(x: Array) -> bool:
     else:
         raise ValueError(f"Unknown type {type(x)}")
 
+
 def stable_acos(x: complex) -> complex:
     """
     Computes a numerically stable arccosine for complex numbers.
@@ -136,7 +142,8 @@ def stable_acos(x: complex) -> complex:
     if np.abs(x) > 1:
         return cmath.acos(x)
     else:
-        return -1j * cmath.log(x + 1j*cmath.sqrt(1 - x*x))
+        return -1j * cmath.log(x + 1j * cmath.sqrt(1 - x * x))
+
 
 def arccos(x: float) -> float:
     """

[…]
