### 2026-05-21 55e4959
initial ruff format of tests
