+++ tests/test_utils.py
@@ -3,8 +3,10 @@ from hypothesis import given, strategies as st
 
 from tmg_hmc.utils import compiled_library_available, arccos
 
+
 def test_compiled_library_installed():
     assert compiled_library_available()
+
 
 @given(st.floats(-1, 1))
 def test_arccos_identity(x):
