+++ tests/test_simple_quadratic_constraints.py
@@ -6,6 +6,7 @@ import scipy.sparse as sp
 
 from tmg_hmc.constraints import SimpleQuadraticConstraint
 from tmg_hmc.gpu_utils import _TORCH_AVAILABLE, torch
+
 if _TORCH_AVAILABLE:
     GPU_AVAILABLE = torch.cuda.is_available()
 else:
@@ -18,21 +19,42 @@ def A_S_vecs(draw, num_vecs=1):
     length = draw(st.integers(min_value=1, max_value=10))
 
     # Generate matrix of shape length x length
-    Amatrix = draw(arrays(dtype=np.float64, shape=(length, length), elements=st.floats(min_value=-1e6, max_value=1e6)))
+    Amatrix = draw(
+        arrays(
+            dtype=np.float64,
+            shape=(length, length),
+            elements=st.floats(min_value=-1e6, max_value=1e6),
+        )
+    )
     Amatrix = (Amatrix + Amatrix.T) / 2  # Ensure the matrix is symmetric
 
     # Generate PSD S matrix
-    Smatrix = draw(arrays(dtype=np.float64, shape=(length, length), elements=st.floats(min_value=-1e6, max_value=1e6)))
+    Smatrix = draw(
+        arrays(
+            dtype=np.float64,
+            shape=(length, length),
+            elements=st.floats(min_value=-1e6, max_value=1e6),
+        )
+    )
     Smatrix = Smatrix @ Smatrix.T  # Make it positive semi-definite
-    Smatrix += np.eye(length) * 1e-6  # Add small value to diagonal for numerical stability
+    Smatrix += (
+        np.eye(length) * 1e-6
+    )  # Add small value to diagonal for numerical stability
 
     # Generate the specified number of vectors with the drawn length
     vectors = [
-        draw(arrays(dtype=np.float64, shape=(length, 1), elements=st.floats(min_value=-1e6, max_value=1e6)))
+        draw(
+            arrays(
+                dtype=np.float64,
+                shape=(length, 1),
+                elements=st.floats(min_value=-1e6, max_value=1e6),
+            )
+        )
         for _ in range(num_vecs)
     ]
 
     return (Amatrix, Smatrix) + tuple(vectors)
+
 
 @st.composite
 def Asparse_S_vecs(draw, num_vecs=1):
@@ -41,22 +63,37 @@ def Asparse_S_vecs(draw, num_vecs=1):
 
     # Generate sparse matrix of shape length x length
     density = draw(st.floats(min_value=0.01, max_value=0.5))
-    Amatrix = sp.random(length, length, density=density, format='coo', dtype=np.float64)
+    Amatrix = sp.random(length, length, density=density, format="coo", dtype=np.float64)
     Amatrix = (Amatrix + Amatrix.T) / 2  # Ensure the matrix is symmetric
     Amatrix = Amatrix.toarray()
 
     # Generate PSD S matrix
-    Smatrix = draw(arrays(dtype=np.float64, shape=(length, length), elements=st.floats(min_value=-1e6, max_value=1e6)))
+    Smatrix = draw(
+        arrays(
+            dtype=np.float64,
+            shape=(length, length),
+            elements=st.floats(min_value=-1e6, max_value=1e6),
+        )
+    )
     Smatrix = Smatrix @ Smatrix.T  # Make it positive semi-definite
-    Smatrix += np.eye(length) * 1e-6  # Add small value to diagonal for numerical stability
+    Smatrix += (
+        np.eye(length) * 1e-6
[…]
