+++ tests/test_sampler.py
@@ -5,34 +5,52 @@ import os
 import numpy as np
 
 from tmg_hmc import TMGSampler
-from tmg_hmc.constraints import LinearConstraint, SimpleQuadraticConstraint, QuadraticConstraint, ProductConstraint
+from tmg_hmc.constraints import (
+    LinearConstraint,
+    SimpleQuadraticConstraint,
+    QuadraticConstraint,
+    ProductConstraint,
+)
 from tmg_hmc.gpu_utils import _TORCH_AVAILABLE, torch
+
 if _TORCH_AVAILABLE:
     GPU_AVAILABLE = torch.cuda.is_available()
 else:
     GPU_AVAILABLE = False
 
+
 @st.composite
 def positive_definite_matrices(draw, min_size=2, max_size=10):
     size = draw(st.integers(min_value=min_size, max_value=max_size))
     # Generate a random matrix
-    A = draw(arrays(dtype=np.float64, shape=(size, size), elements=st.floats(-1e-6, 1e-6)))
+    A = draw(
+        arrays(dtype=np.float64, shape=(size, size), elements=st.floats(-1e-6, 1e-6))
+    )
     # Make it symmetric positive definite
-    pd_matrix = A.T @ A + np.eye(size) * 1e-6  # Add small value to diagonal for numerical stability
+    pd_matrix = (
+        A.T @ A + np.eye(size) * 1e-6
+    )  # Add small value to diagonal for numerical stability
     return pd_matrix
+
 
 @st.composite
 def square_matrix(draw, min_size=2, max_size=10):
     size = draw(st.integers(min_value=min_size, max_value=max_size))
-    mat = draw(arrays(dtype=np.float64, shape=(size, size), elements=st.floats(-1e6, 1e6)))
+    mat = draw(
+        arrays(dtype=np.float64, shape=(size, size), elements=st.floats(-1e6, 1e6))
+    )
     return mat
+
 
 @st.composite
 def symmetric_matrix(draw, min_size=2, max_size=10):
     size = draw(st.integers(min_value=min_size, max_value=max_size))
-    mat = draw(arrays(dtype=np.float64, shape=(size, size), elements=st.floats(-1e6, 1e6)))
+    mat = draw(
+        arrays(dtype=np.float64, shape=(size, size), elements=st.floats(-1e6, 1e6))
+    )
     sym_mat = (mat + mat.T) / 2
     return sym_mat
+
 
 @st.composite
 def vector(draw, min_size=2, max_size=10):
@@ -40,13 +58,17 @@ def vector(draw, min_size=2, max_size=10):
     vec = draw(arrays(dtype=np.float64, shape=(size, 1), elements=st.floats(-1e6, 1e6)))
     return vec
 
+
 @st.composite
 def sym_mat_vec(draw, min_size=2, max_size=10):
     size = draw(st.integers(min_value=min_size, max_value=max_size))
-    mat = draw(arrays(dtype=np.float64, shape=(size, size), elements=st.floats(-1e6, 1e6)))
+    mat = draw(
+        arrays(dtype=np.float64, shape=(size, size), elements=st.floats(-1e6, 1e6))
+    )
     sym_mat = (mat + mat.T) / 2
     vec = draw(arrays(dtype=np.float64, shape=(size, 1), elements=st.floats(-1e6, 1e6)))
     return sym_mat, vec
+
 
 @given(positive_definite_matrices())
 def test_sampler_initialization_errors(pd_matrix):
@@ -61,18 +83,23 @@ def test_sampler_initialization_errors(pd_matrix):
         TMGSampler(Sigma_half=pd_matrix[:-1, :])  # Invalid shape
 
[…]
