### 2026-05-21 83daa9e
initial ruff formatting and checking of base directory
