### 2026-05-21 4d05565
updated psd generation to be more numerically stable
