+++ tests/test_quadratic_constraints.py
@@ -23,7 +23,7 @@ def A_S_vecs(draw, num_vecs=2):
         arrays(
             dtype=np.float64,
             shape=(length, length),
-            elements=st.floats(min_value=-1e6, max_value=1e6),
+            elements=st.floats(min_value=-1e3, max_value=1e3),
         )
     )
     Amatrix = (Amatrix + Amatrix.T) / 2  # Ensure the matrix is symmetric
@@ -33,7 +33,7 @@ def A_S_vecs(draw, num_vecs=2):
         arrays(
             dtype=np.float64,
             shape=(length, length),
-            elements=st.floats(min_value=-1e6, max_value=1e6),
+            elements=st.floats(min_value=-1e3, max_value=1e3),
         )
     )
     Smatrix = Smatrix @ Smatrix.T  # Make it positive semi-definite
@@ -47,7 +47,7 @@ def A_S_vecs(draw, num_vecs=2):
             arrays(
                 dtype=np.float64,
                 shape=(length, 1),
-                elements=st.floats(min_value=-1e6, max_value=1e6),
+                elements=st.floats(min_value=-1e3, max_value=1e3),
             )
         )
         for _ in range(num_vecs)
@@ -72,7 +72,7 @@ def Asparse_S_vecs(draw, num_vecs=2):
         arrays(
             dtype=np.float64,
             shape=(length, length),
-            elements=st.floats(min_value=-1e6, max_value=1e6),
+            elements=st.floats(min_value=-1e3, max_value=1e3),
         )
     )
     Smatrix = Smatrix @ Smatrix.T  # Make it positive semi-definite
@@ -86,7 +86,7 @@ def Asparse_S_vecs(draw, num_vecs=2):
             arrays(
                 dtype=np.float64,
                 shape=(length, 1),
-                elements=st.floats(min_value=-1e6, max_value=1e6),
+                elements=st.floats(min_value=-1e3, max_value=1e3),
             )
         )
         for _ in range(num_vecs)
@@ -95,7 +95,7 @@ def Asparse_S_vecs(draw, num_vecs=2):
     return (Amatrix, Smatrix) + tuple(vectors)
 
 
-@given(A_S_vecs(), st.floats(min_value=-1e6, max_value=1e6))
+@given(A_S_vecs(), st.floats(min_value=-1e3, max_value=1e3))
 def test_quadratic_constraint_value(input_lists, c):
     A, S, f, x = input_lists
     assume(not np.allclose(A, np.zeros_like(A)))  # Ensure A is not the zero matrix
@@ -111,7 +111,7 @@ def test_quadratic_constraint_value(input_lists, c):
 
 @pytest.mark.gpu
 @pytest.mark.skipif(not GPU_AVAILABLE, reason="GPU not available")
-@given(A_S_vecs(), st.floats(min_value=-1e6, max_value=1e6))
+@given(A_S_vecs(), st.floats(min_value=-1e3, max_value=1e3))
 def test_quadratic_constraint_value_gpu(input_lists, c):
     A, S, f, x = input_lists
     assume(not np.allclose(A, np.zeros_like(A)))  # Ensure A is not the zero matrix
@@ -129,7 +129,7 @@ def test_quadratic_constraint_value_gpu(input_lists, c):
     assert np.isclose(val, expected_val.cpu().item())
 
 
-@given(Asparse_S_vecs(), st.floats(min_value=-1e6, max_value=1e6))
[…]

[…]
