### 2026-05-21 e650556
added clang and cppcheck and updated cpp files
