+++ src/tmg_hmc/quad_solns.h
@@ -1,7 +1,8 @@
-#include <pybind11/pybind11.h>
 #include <pybind11/numpy.h>
-#include <complex>
+#include <pybind11/pybind11.h>
+
 #include <cmath>
+#include <complex>
 #include <vector>
 // #include <iostream>
 
@@ -20,7 +21,8 @@ double arccos(std::complex<long double> z) {
     double real_part = static_cast<double>(std::real(val));
     double imag_part = static_cast<double>(std::imag(val));
     if (std::abs(imag_part) > 1e-1) {
-        return std::numeric_limits<double>::quiet_NaN();  // Return NaN for significant imaginary parts
+        return std::numeric_limits<double>::quiet_NaN();  // Return NaN for significant
+                                                          // imaginary parts
     }
     return real_part;
 }
@@ -36,44 +38,64 @@ inline std::complex<double> operator*(const std::complex<double>& lhs, int rhs) 
 }
 
 // Allow same for complex<long double>
-inline std::complex<long double> operator*(int lhs, const std::complex<long double>& rhs) {
+inline std::complex<long double> operator*(int lhs,
+                                           const std::complex<long double>& rhs) {
     return static_cast<long double>(lhs) * rhs;
 }
 
 // Allow complex<long double> * int
-inline std::complex<long double> operator*(const std::complex<long double>& lhs, int rhs) {
+inline std::complex<long double> operator*(const std::complex<long double>& lhs,
+                                           int rhs) {
     return lhs * static_cast<long double>(rhs);
 }
 
 // Allow complex<long double> * double
-inline std::complex<long double> operator*(const std::complex<long double>& lhs, double rhs) {
+inline std::complex<long double> operator*(const std::complex<long double>& lhs,
+                                           double rhs) {
     return lhs * static_cast<long double>(rhs);
 }
 
 // Allow double * complex<long double>
-inline std::complex<long double> operator*(double lhs, const std::complex<long double>& rhs) {
+inline std::complex<long double> operator*(double lhs,
+                                           const std::complex<long double>& rhs) {
     return static_cast<long double>(lhs) * rhs;
 }
 
 // Allow complex<long double> / double
-inline std::complex<long double> operator/(const std::complex<long double>& lhs, double rhs) {
+inline std::complex<long double> operator/(const std::complex<long double>& lhs,
+                                           double rhs) {
     return lhs / static_cast<long double>(rhs);
 }
 
-
-
 double* calc_all_solutions(double q1, double q2, double q3, double q4, double q5);
-double soln1(std::complex<double> q1in, std::complex<double> q2in, std::complex<double> q3in, std::complex<double> q4in, std::complex<double> q5in);
-double soln2(std::complex<double> q1in, std::complex<double> q2in, std::complex<double> q3in, std::complex<double> q4in, std::complex<double> q5in);
-double soln3(std::complex<double> q1in, std::complex<double> q2in, std::complex<double> q3in, std::complex<double> q4in, std::complex<double> q5in);
-double soln4(std::complex<double> q1in, std::complex<double> q2in, std::complex<double> q3in, std::complex<double> q4in, std::complex<double> q5in);
-double soln5(std::complex<double> q1in, std::complex<double> q2in, std::complex<double> q3in, std::complex<double> q4in, std::complex<double> q5in);
-double soln6(std::complex<double> q1in, std::complex<double> q2in, std::complex<double> q3in, std::complex<double> q4in, std::complex<double> q5in);
[…]
