### 2026-05-21 9fd8f0b
added pre-commit instructions
