### 2026-05-21 dabd848
initial constraints module split up
