+++ src/tmg_hmc/constraints/base.py
@@ -1,0 +1,361 @@
+from __future__ import annotations
+import numpy as np
+from typing import Tuple, Protocol
+from tmg_hmc.utils import Array, get_sparse_elements
+from tmg_hmc.gpu_utils import torch, Tensor
+
+pis = np.array([-1, 0, 1]) * np.pi
+eps = 1e-12
+
+
+class Constraint(Protocol):
+    """
+    Abstract base class for constraints
+    """
+
+    def value(self, x: Array) -> float:
+        """
+        Compute the value of the constraint at x
+        """
+        pass
+
+    def is_satisfied(self, x: Array) -> bool:
+        """
+        Check if the constraint is satisfied at x
+
+        Parameters
+        ----------
+        x : Array
+            Point to evaluate the constraint at
+        Returns
+        -------
+        bool
+            True if the constraint is satisfied, False otherwise
+        """
+        return self.value(x) >= 0
+
+    def is_zero(self, x: Array) -> Tuple[bool, bool]:
+        """
+        Check if the constraint is zero at x
+
+        Parameters
+        ----------
+        x : Array
+            Point to evaluate the constraint at
+        Returns
+        -------
+        Tuple[bool, bool]
+            (is_strictly_zero, is_approximately_zero)
+        """
+        val = self.value(x)
+        return np.isclose(val, 0), np.isclose(val, 0, atol=1e-2)
+
+    def compute_q(self, a: Array, b: Array) -> Tuple[float, ...]:
+        """
+        Compute the coefficients of the constraint equation along the trajectory defined by a and b
+        """
+        pass
+
+    def hit_time(self, a: Array, b: Array) -> Array:
+        """
+        Compute the hit time of the constraint along the trajectory defined by a and b
+        """
+        pass
+
+    def normal(self, x: Array) -> Array:
+        """
+        Compute the normal vector of the constraint at x
+        """
+        pass
+
+    def reflect(self, x: Array, xdot: Array) -> Array:
+        """
+        Reflect the velocity xdot at the constraint surface defined by x
+
+        Parameters
+        ----------
+        x : Array
+            Point on the constraint surface
+        xdot : Array
+            Velocity to be reflected
+
+        Returns
+        -------
+        Array
+            Reflected velocity
+        """
+        f = self.normal(x)
+        if isinstance(xdot, Tensor):
+            norm = torch.sqrt(f.T @ f)
+        else:
+            norm = np.sqrt(f.T @ f)
+        f = f / norm
+        return xdot - 2 * (f.T @ xdot) * f
+
+    def serialize(self) -> dict:
+        """
+        Serialize the constraint to a dictionary
+
+        Returns
+        -------
+        dict
+            Dictionary representation of the constraint
+        """
+        d = self.__dict__.copy()
+
+        # For sparse constraints, ensure we save S directly
+        # and remove the individual row/column vectors that cause reconstruction issues
+        if "sparse" in d and d["sparse"]:
+            # Keep S if it exists
+            if "S" not in d and hasattr(self, "S"):
+                d["S"] = self.S
+            # Remove problematic sparse reconstruction data
+            keys_to_remove = ["s_rows", "s_cols", "row_data", "col_data"]
+            for key in keys_to_remove:
+                if key in d:
+                    del d[key]
+
+        # Convert tensors to CPU
+        for k, v in d.items():
+            if isinstance(v, Tensor):
+                d[k] = v.cpu()
+
+        d["type"] = self.__class__.__name__
+        return d
+
+    # @classmethod
+    # def deserialize(cls, d: dict, gpu: bool) -> Constraint:
+    #     """
+    #     Deserialize the constraint from a dictionary
+
+    #     Parameters
+    #     ----------
+    #     d : dict
[…]
