+++ src/tmg_hmc/constraints/quadratic.py
@@ -1,0 +1,524 @@
+from __future__ import annotations
+import numpy as np
+from typing import Tuple
+from tmg_hmc.utils import Array, Sparse, to_scalar
+from tmg_hmc.quad_solns import soln1, soln2, soln3, soln4, soln5, soln6, soln7, soln8
+from tmg_hmc.gpu_utils import Tensor, _TORCH_AVAILABLE
+from tmg_hmc.compiled import calc_all_solutions
+from tmg_hmc.constraints.base import BaseQuadraticConstraint, pis, eps
+import warnings
+
+
+class SimpleQuadraticConstraint(BaseQuadraticConstraint):
+    """
+    Constraint of the form x^T A x + c >= 0
+    """
+
+    def __init__(self, A: Array, c: float, S: Array, sparse: bool = False):
+        """
+        Parameters
+        ----------
+        A : Array
+            Quadratic coefficient matrix
+        c : float
+            Constant term
+        S : Array
+            Transformation matrix given by the Symmetric Sqrt of the Mass matrix
+        sparse : bool, optional
+            Whether to use sparse matrix computations, by default False
+
+        Notes
+        -----
+        If A is a sparse matrix, sparse computations are used regardless of the
+        sparse parameter.
+        """
+        self.c = c
+        if isinstance(A, Sparse):
+            sparse = True
+        self.sparse = sparse
+        if sparse:
+            self._setup_values_sparse(A, S)
+        else:
+            self._setup_values(A, S)
+
+    @classmethod
+    def build_from_dict(cls, d: dict, gpu: bool) -> SimpleQuadraticConstraint:
+        """
+        Build a SimpleQuadraticConstraint from a dictionary representation
+
+        Parameters
+        ----------
+        d : dict
+            Dictionary representation of the constraint
+        gpu : bool
+            Whether to load tensors onto the GPU
+
+        Returns
+        -------
+        SimpleQuadraticConstraint
+            The constructed constraint
+        """
+        if gpu and not _TORCH_AVAILABLE:
+            gpu = False
+            warnings.warn(
+                "GPU requested but PyTorch is not available. Loading on CPU instead."
+            )
+        sparse = d["sparse"]
+        A = d["A_orig"]
+        c = d["c"]
+        S = d.get("S", None)
+
+        # Move to GPU if requested
+        if gpu:
+            if isinstance(S, Tensor):
+                S = S.cuda()
+            if isinstance(A, Tensor):
+                A = A.cuda()
+
+        return cls(A, c, S, sparse)
+
+    def value_(self, x: Array) -> float:
+        """
+        Compute the value of the constraint at x using dense matrix computations
+
+        Parameters
+        ----------
+        x : Array
+            Point to evaluate the constraint at
+
+        Returns
+        -------
+        float
+            Value of the constraint at x given by x^T A x + c
+        """
+        return to_scalar(x.T @ self.A @ x + self.c)
+
+    def value_sparse(self, x: Array) -> float:
+        """
+        Compute the value of the constraint at x using sparse matrix computations
+
+        Parameters
+        ----------
+        x : Array
+            Point to evaluate the constraint at
+
+        Returns
+        -------
+        float
+            Value of the constraint at x given by x^T A x + c
+        """
+        return to_scalar(x.T @ self.A_dot_x(x) + self.c)
+
+    def normal_(self, x: Array) -> Array:
+        """
+        Compute the normal vector at x using dense matrix computations
+
+        Parameters
+        ----------
+        x : Array
+            Point to evaluate the normal vector at
+
+        Returns
+        -------
+        Array
+            Normal vector at x given by 2 * A @ x
+        """
[…]

[…]
