### 2026-05-21 6c7d6ca
added registry pattern to fix deserialization issues
