### 2026-05-21 46f0c0d
added build from dict method to linear constraint to follow the deserialize pattern
