### 2026-05-21 b771717
added mypy
