+++ pyproject.toml
@@ -42,6 +42,7 @@ test = [
     "pytest-cov",
     "hypothesis>=6.100",
     "ruff",
+    "mypy",
     "pre-commit",
 ]
 
