### 2026-05-22 b637074
ran and fixed mypy type issues
