+++ src/tmg_hmc/sampler.py
@@ -22,19 +22,19 @@ class TMGSampler:
 
     def __init__(
         self,
-        mu: Array = None,
-        Sigma: Array = None,
+        mu: Array | None = None,
+        Sigma: Array | None = None,
         T: float = np.pi / 2,
         gpu: bool = False,
         *,
-        Sigma_half: Array = None,
+        Sigma_half: Array | None = None,
     ) -> None:
         """
         Parameters
         ----------
         mu : Array, optional
             Mean vector of the Gaussian distribution. If None, defaults to zero vector.
-        Sigma : Array
+        Sigma : Array, optional
             Covariance matrix of the Gaussian distribution. Must be positive semi-definite.
             Do not provide if Sigma_half is given.
         T : float, optional
@@ -45,22 +45,23 @@ class TMGSampler:
             Matrix such that Sigma_half @ Sigma_half.T = Sigma.
             If provided, Sigma is not needed.
         """
-        if Sigma is None and Sigma_half is None:
+        if Sigma_half is not None:
+            self.dim = len(Sigma_half)
+            self._setup_sigma_half(Sigma_half)
+        elif Sigma is not None:
+            self.dim = len(Sigma)
+            self._setup_sigma(Sigma)
+        else:
             raise ValueError("Must provide either Sigma or Sigma_half")
-        self.dim = len(Sigma) if Sigma is not None else len(Sigma_half)
         if mu is None:
             mu = np.zeros(self.dim)
         self.mu = mu.reshape(self.dim, 1)
         self.T = T
-        self.constraints = []
+        self.constraints: list[Constraint] = []
         self.constraint_violations = 0
         self.gpu = gpu
-        self.x = None
+        self.x: Array | None = None
 
-        if Sigma_half is not None:
-            self._setup_sigma_half(Sigma_half)
-        else:
-            self._setup_sigma(Sigma)
         if self.gpu:
             self.mu = torch.tensor(self.mu).cuda()
 
@@ -96,10 +97,10 @@ class TMGSampler:
         if self.gpu:
             Sigma = torch.tensor(Sigma).cuda()
             s, V = torch.linalg.eigh(Sigma)
-            all_positive = torch.all(s >= 0)
+            all_positive = bool(torch.all(s >= 0))
         else:
             s, V = np.linalg.eigh(Sigma)
-            all_positive = np.all(s >= 0)
+            all_positive = bool(np.all(s >= 0))
         if not all_positive:
             min_eig = torch.min(s) if self.gpu else np.min(s)
             if abs(min_eig) < 1e-10:
@@ -132,10 +133,10 @@ class TMGSampler:
         if self.gpu:
             Sigma_half = torch.tensor(Sigma_half).cuda()
             s, V = torch.linalg.eigh(Sigma_half)
-            all_positive = torch.all(s >= 0)
+            all_positive = bool(torch.all(s >= 0))
         else:
             s, V = np.linalg.eigh(Sigma_half)
-            all_positive = np.all(s >= 0)
+            all_positive = bool(np.all(s >= 0))
         if not all_positive:
             min_eig = torch.min(s) if self.gpu else np.min(s)
             if abs(min_eig) < 1e-10:
@@ -150,8 +151,8 @@ class TMGSampler:
     def _build_constraint(
         self,
         *,
-        A: Array = None,
-        f: Array = None,
+        A: Array | None = None,
+        f: Array | None = None,
         c: float = 0.0,
         sparse: bool = True,
         compiled: bool = True,
@@ -201,9 +202,9 @@ class TMGSampler:
                 f = torch.tensor(f).cuda()
 
         if (A is not None) and sparse:
-            A = sparsify(A)
+            A = sparsify(A) # type: ignore[assignment]
         if (f is not None) and sparse:
[…]
