+++ pyproject.toml
@@ -23,7 +23,8 @@ classifiers = [
 ]
 dependencies = [
     "numpy",
-    "scipy"
+    "scipy",
+    "scipy-stubs"
 ]
 
 [project.optional-dependencies]
@@ -82,3 +83,8 @@ repair-wheel-command = ""
 [tool.cibuildwheel.windows]
 before-build = "pip install delvewheel"
 repair-wheel-command = "delvewheel repair -w {dest_dir} {wheel}"
+
+[tool.mypy]
+[[tool.mypy.overrides]]
+module = "tmg_hmc.compiled"
+ignore_missing_imports = true

[…]
