### 2026-05-22 49484c4
added detailed documentation to notebooks
