### 2026-05-22 f7b9357
initial sphinx build
