+++ docs/source/conf.py
@@ -1,0 +1,47 @@
+# Configuration file for the Sphinx documentation builder.
+
+import os
+import sys
+
+sys.path.insert(0, os.path.abspath("../../src"))
+
+# -- Project information -----------------------------------------------------
+project = "tmg-hmc"
+copyright = "2026, Erik A. Bensen, Mikael Kuusela"
+author = "Erik A. Bensen, Mikael Kuusela"
+release = "1.0.3"
+
+# -- General configuration ---------------------------------------------------
+extensions = [
+    "sphinx.ext.autodoc",  # pulls docstrings automatically
+    "sphinx.ext.napoleon",  # parses NumPy-style docstrings
+    "sphinx.ext.mathjax",  # renders LaTeX math
+    "sphinx.ext.viewcode",  # adds links to source code
+    "sphinx_autodoc_typehints",  # pulls type annotations into docs
+]
+
+templates_path = ["_templates"]
+exclude_patterns = ["_build"]
+
+# Napoleon settings - tell Sphinx your docstrings are NumPy style
+napoleon_google_docstring = False
+napoleon_numpy_docstring = True
+
+# Autodoc settings
+autodoc_member_order = "bysource"  # document members in source order
+autodoc_typehints = "description"  # put type hints in the description
+
+# -- Options for HTML output -------------------------------------------------
+html_theme = "sphinx_rtd_theme"
+html_static_path = ["_static"]
+
+mathjax_path = "https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js"
+
+mathjax3_config = {
+    "tex": {
+        "inlineMath": [["\\(", "\\)"]],
+        "displayMath": [["\\[", "\\]"]],
+        "packages": {"[+]": ["boldsymbol"]},
+    },
+    "loader": {"load": ["[tex]/boldsymbol"]},
+}
