+++ docs/source/api.rst
@@ -1,0 +1,38 @@
+API Reference
+=============
+
+Sampler
+-------
+
+.. autoclass:: tmg_hmc.TMGSampler
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Constraints
+-----------
+
+.. autoclass:: tmg_hmc.constraints.Constraint
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. autoclass:: tmg_hmc.constraints.LinearConstraint
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. autoclass:: tmg_hmc.constraints.SimpleQuadraticConstraint
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. autoclass:: tmg_hmc.constraints.QuadraticConstraint
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. autoclass:: tmg_hmc.constraints.ProductConstraint
+   :members:
+   :undoc-members:
+   :show-inheritance:
