### 2026-05-22 7e49b12
changed quadratic linear term to f for naming consistency
