### 2026-05-22 b3a30e6
updated docstrings to use bold for vectors and capitals for matrices
