### 2026-05-25 d29a951
added more mathjax documentation
