### 2026-05-26 7988c1c
added doc dependencies
