### 2026-05-26 9bbe40f
added updated api docs and examples
