+++ docs/source/citing.rst
@@ -1,0 +1,54 @@
+Citing tmg-hmc
+==============
+
+If you use this package in your research, please cite both the software and the
+underlying methodology.
+
+Software
+--------
+
+.. code-block:: text
+
+   Bensen, E. A., & Kuusela, M. (2026). tmg_hmc: A Python package for Exact HMC
+   Sampling for Truncated Multivariate Gaussians with Linear and Quadratic
+   Constraints. Journal of Open Source Software. [In Review]
+
+BibTeX:
+
+.. code-block:: bibtex
+
+   @article{Bensen2026tmghmc,
+     title={tmg\_hmc: A Python package for Exact HMC Sampling for Truncated
+            Multivariate Gaussians with Linear and Quadratic Constraints},
+     author={Bensen, Erik A. and Kuusela, Mikael},
+     journal={Journal of Open Source Software},
+     year={2026},
+     note={[In Review]}
+   }
+
+Methodology
+-----------
+
+The algorithm implemented in this package is described in:
+
+.. code-block:: text
+
+   Pakman, A., & Paninski, L. (2014). Exact Hamiltonian Monte Carlo for Truncated
+   Multivariate Gaussians. Journal of Computational and Graphical Statistics,
+   23(2), 518-542. https://doi.org/10.1080/10618600.2013.788448
+
+BibTeX:
+
+.. code-block:: bibtex
+
+   @article{PakmanPaninski2014,
+     title={Exact Hamiltonian Monte Carlo for Truncated Multivariate Gaussians},
+     author={Pakman, Ari and Paninski, Liam},
+     journal={Journal of Computational and Graphical Statistics},
+     volume={23},
+     number={2},
+     pages={518--542},
+     year={2014},
+     publisher={Taylor \& Francis},
+     doi={10.1080/10618600.2013.788448}
+   }
