+++ docs/source/examples.rst
@@ -1,0 +1,10 @@
+Examples
+========
+
+.. toctree::
+   :maxdepth: 1
+
+   examples/single_constraint_examples
+   examples/multi-constraint_examples
+   examples/product-constraint_examples
+   examples/gaussian_process_examples
