+++ pyproject.toml
@@ -51,6 +51,7 @@ docs = [
     "sphinx",
     "sphinx-rtd-theme",
     "sphinx-autodoc-typehints",
+    "nbsphinx",
 ]
 
 
