### 2026-05-26 dab574a
updated README to point to readthedocs documentation
