### 2026-05-26 16d5562
moved documentation above readme
