### 2026-06-05 6f6a23f
Update API documentation [skip ci]
