### 2026-06-05 7d28e30
bump version to 1.0.4
