+++ src/tmg_hmc/__init__.py
@@ -1,5 +1,5 @@
 from .sampler import TMGSampler
 
-__version__ = "1.0.3"
+__version__ = "1.0.4"
 
 __all__ = ["TMGSampler"]
