+++ API_DOCS.md
@@ -1,640 +1,0 @@
-# API Reference
-
-> **Package:** `tmg_hmc` v1.0.3
->
-> This package implements exact HMC sampling for truncated multivariate gaussians with quadratic constraints.
-
-## `TMGSampler` (class)
-
-```python
-TMGSampler(mu: Array | None = None, Sigma: Array | None = None, T: float | None = None, gpu: bool = False, Sigma_half: Array | None = None)
-```
-
-Hamiltonian Monte Carlo sampler for Multivariate Gaussian distributions<br>
-with linear and quadratic constraints.
-
-**Constructor:**
-
-Parameters<br>
-----------<br>
-mu : Array, optional<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Mean vector of the Gaussian distribution. If None, defaults to zero vector.<br>
-Sigma : Array, optional<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Covariance matrix of the Gaussian distribution. Must be positive semi-definite.<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Do not provide if Sigma_half is given.<br>
-T : float, optional<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Integration time for the Hamiltonian dynamics. Default is :math:`\pi/2`.<br>
-gpu : bool, optional<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Whether to use GPU acceleration with PyTorch. Default is False.<br>
-Sigma_half : Array, optional<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Matrix such that :math:`S S^T = \Sigma`.<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;If provided, Sigma is not needed.
-
-### Methods
-
-### `add_constraint`
-
-```python
-add_constraint(A: Array | None = None, f: Array | None = None, c: float = 0.0, sparse: bool = True, compiled: bool = True) -> None
-```
-
-Adds a constraint to the sampler.
-
-.. math::
-
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;\mathbf{x}^T A \mathbf{x} + \mathbf{f}^T \mathbf{x} + c \geq 0
-
-Parameters<br>
-----------<br>
-A : Array, optional<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Quadratic term matrix, defaults to the zero matrix if not provided.<br>
-f : Array, optional<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Linear term vector, defaults to the zero vector if not provided.<br>
-c : float, optional<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Constant term. Default is 0.0.<br>
-sparse : bool, optional<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Whether to store A and f in sparse format. Default is True.<br>
-compiled : bool, optional<br>
[…]
