+++ pyproject.toml
@@ -8,7 +8,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "tmg_hmc"
-version = "1.0.3"
+version = "1.0.4"
 authors = [
   { name="Erik A. Bensen", email="ebensen@andrew.cmu.edu" },
   { name="Mikael Kuusela", email="mkuusela@andrew.cmu.edu" },
