+++ .github/workflows/generate-docs.yml
@@ -1,412 +1,0 @@
-name: Generate API Documentation
-
-on:
-  push:
-    branches: [ main, dev ]
-    paths:
-      - 'src/**/*.py'
-      - 'src/**/*.h'
-
-permissions:
-  contents: write
-
-jobs:
-  generate-docs:
-    runs-on: ubuntu-latest
-
-    steps:
-    - name: Checkout repository
-      uses: actions/checkout@v4
-
-    - name: Skip if only __init__.py changed
-      id: skip
-      run: |
-        # List all changed files in this push
-        changed=$(git diff --name-only ${{ github.event.before }} ${{ github.sha }} | grep -v '__init__.py' || true)
-        if [ -z "$changed" ]; then
-          echo "Only __init__.py files changed. Skipping documentation generation."
-          exit 0
-        fi
-    
-    - name: Set up Python
-      uses: actions/setup-python@v5
-      with:
-        python-version: '3.10'
-        cache: 'pip'
-    
-    - name: Cache pip packages
-      uses: actions/cache@v4
-      with:
-        path: ~/.cache/pip
-        key: ${{ runner.os }}-pip-${{ hashFiles('pyproject.toml') }}
-        restore-keys: |
-          ${{ runner.os }}-pip-
-    
-    - name: Install build dependencies
-      run: |
-        sudo apt-get update
-        sudo apt-get install -y build-essential g++ make
-    
-    - name: Install package (will compile C++ automatically)
-      run: |
-        pip install -e .
-    
-    - name: Generate API documentation
-      run: |
-        python << 'EOF'
-        import inspect
-        import importlib
-        import pkgutil
-        import tmg_hmc
-        import ast
-        import re
-        from pathlib import Path
-
-        # ============================================================
-        # CONFIGURE MODULES HERE - Edit this list to control order
-        # Set to None to auto-discover all modules
-        # ============================================================
-        MODULES_TO_DOCUMENT = [
-            'tmg_hmc',
-            'tmg_hmc.constraints',
-            'tmg_hmc.utils',
-            'tmg_hmc.compiled',
-            'tmg_hmc.quad_solns',
-        ]
-        # ============================================================
-
-        def clean_docstring(docstring):
-            """Clean and format docstring for markdown display."""
-            if not docstring:
-                return None
-            
-            cleaned = inspect.cleandoc(docstring)
-            lines = cleaned.split('\n')
-            result = []
-            
-            for i, line in enumerate(lines):
-                if line.strip() == '':
-                    result.append('')
-                else:
-                    stripped = line.lstrip()
-                    leading_spaces = len(line) - len(stripped)
-                    
-                    if leading_spaces > 0:
-                        indent = '&nbsp;' * (leading_spaces * 2)
-                        line_content = f'{indent}{stripped}'
-                    else:
-                        line_content = stripped
-                    
-                    if i < len(lines) - 1 and lines[i + 1].strip() != '':
-                        line_content += '<br>'
-                    
-                    result.append(line_content)
-            
-            return '\n'.join(result)
-
-        def get_source_defaults(func):
-            """Extract default values from source code, preserving expressions."""
-            try:
-                source = inspect.getsource(func)
-                tree = ast.parse(source)
-                
-                for node in ast.walk(tree):
-                    if isinstance(node, (ast.FunctionDef, ast.AsyncFunctionDef)):
-                        defaults_dict = {}
-                        args = node.args
-                        
-                        # Handle positional defaults
-                        num_defaults = len(args.defaults)
-                        if num_defaults > 0:
-                            default_offset = len(args.args) - num_defaults
-                            for i, default in enumerate(args.defaults):
-                                arg_name = args.args[default_offset + i].arg
-                                defaults_dict[arg_name] = ast.unparse(default)
[…]
