### 2026-08-24 27ca828
Revise citation for tmg_hmc package
Updated citation to include a URL for the tmg_hmc package and removed 'In Review' note.
