### 2026-08-24 8227abf
Fix year in BibTeX citation for tmg_hmc
Corrected the year in the BibTeX entry for the tmg_hmc package.
