### 2026-08-24 5f7ca33
Change BibTeX entry type from article to misc
Updated BibTeX entry from article to misc type.
